+++ examples/hypervisor_managers.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 ###
-# (C) Copyright [2019] Hewlett Packard Enterprise Development LP
+# (C) Copyright [2020] Hewlett Packard Enterprise Development LP
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -15,6 +15,8 @@
 # limitations under the License.
 ###
 
+from pprint import pprint
+
 from config_loader import try_load_from_file
 from hpOneView.oneview_client import OneViewClient
 
@@ -26,7 +28,7 @@ config = {
     }
 }
 
-hypervisor_manager_information = {
+options = {
     "type": "HypervisorManagerV2",
     "name": "172.18.13.11",
     "displayName": "vcenter",
@@ -39,44 +41,53 @@ hypervisor_manager_information = {
 # Try load config from a file (if there is a config file)
 config = try_load_from_file(config)
 oneview_client = OneViewClient.config()
+hypervisor_managers = oneview_client.hypervisor_managers
 
-# Add a hypervisor manager
-hypervisor_manager_added = oneview_client.hypervisor_managers.add(hypervisor_manager_information)
-print('Added hypervisor manager "{name}" successfully\n'.format(**hypervisor_manager_added))
+# Find recently created hypervisor manager by name
+print("\nGet Hypervisor Manager by name")
+hypervisor_manager = hypervisor_managers.get_by_name(options['name'])
 
-# Retrieve hypervisor manager by URI
-hypervisor_manager = oneview_client.hypervisor_managers.get(hypervisor_manager_added['uri'])
-print('Get hypervisor manager by URI "{uri}", retrieved "{name}" successfully\n'.format(**hypervisor_manager))
+if hypervisor_manager:
+    print("\nFound hypervisor-manager by name: {}.\n  uri = {}".format(hypervisor_manager.data['name'], hypervisor_manager.data['uri']))
+else:
+    # Create a HypervisorManager with the options provided
+    hypervisor_manager = hypervisor_managers.create(data=options)
+    print("\nCreated a hypervisor-manager with name: {}.\n  uri = {}".format(hypervisor_manager.data['name'], hypervisor_manager.data['uri']))
 
-# Retrieve hypervisor manager by FILTER
-hypervisor_manager_list = oneview_client.hypervisor_managers.get_by('hypervisorType', 'Vmware')
-print('Get hypervisor manager by FILTER "{hypervisorType}", retrieved "{name}" successfully\n'.format(**hypervisor_manager_list[0]))
+# Get all, with defaults
+print("\nGet all hypervisor managers")
+hyp_managers_all = hypervisor_manager.get_all()
+for hyp in hyp_managers_all:
+    print('  - {}'.format(hyp['name']))
 
-# Update the hypervisor manager
-hypervisor_manager['displayName'] = "New hypervisor"
-hypervisor_manager = oneview_client.hypervisor_managers.update(hypervisor_manager)
-print('Hypervisor manager "{displayName}" updated successfully\n'.format(**hypervisor_manager))
+# Get the first 10 records
+print("\nGet the first ten hypervisor managers")
+hyp_mgrs_top_ten = hypervisor_managers.get_all(0, 10)
+for hyp in hyp_mgrs_top_ten:
+    print('  - {}'.format(hyp['name']))
 
-# Retrieve hypervisor manager by NAME
-hypervisor_manager = oneview_client.hypervisor_managers.get_by_name(hypervisor_manager['name'])
-print('Get hypervisor manager by NAME "{name}", retrieved "{uri}" successfully\n'.format(**hypervisor_manager))
+# Filter by hypervisor type
+print("\nGet all hypervisor managers filtering by hypervisor type")
+hyp_mgrs_filtered = hypervisor_managers.get_all(filter="\"'hypervisorType'='Vmware'\"")
+for hyp in hyp_mgrs_filtered:
+    print("Hypervisor with type 'Vmware'  - {}".format(hyp['name']))
 
-# Update the hypervisor manager forcefully
-hypervisor_manager['displayName'] = "Update hypervisor force"
-hypervisor_manager = oneview_client.hypervisor_managers.update(hypervisor_manager, force=True)
-print('Hypervisor manager "{displayName}" updated forcefully\n'.format(**hypervisor_manager))
+# Get all sorting by name descending
+print("\nGet all hypervisor managers sorting by name")
+hyp_mgrs_sorted = hypervisor_managers.get_all(sort='name:descending')
+pprint(hyp_mgrs_sorted)
 
-# Get all hypervisor managers
-print("Get all hypervisor managers:")
-hypervisor_manager_all = oneview_client.hypervisor_managers.get_all()
-for hyp_mgr in hypervisor_manager_all:
-    print(" - " + hyp_mgr['name'])
+# Get by uri
+print("\nGet a hypervisor managers by uri")
+hyp_mgrs_by_uri = hypervisor_managers.get_by_uri(hypervisor_manager.data['uri'])
+pprint(hyp_mgrs_by_uri.data)
 
-# Remove added hypervisor manager
-oneview_client.hypervisor_managers.delete(hypervisor_manager)
-print('Successfully removed the hypervisor manager "{name}"\n'.format(**hypervisor_manager))
+# Update display name of recently created hypervisor manager
+data_to_update = {'displayName': 'Updated vcenter'}
+hypervisor_manager.update(data=data_to_update)
+print("\nUpdated hypervisor manager {} successfully.\n  uri = {}".format(hypervisor_manager.data['name'], hypervisor_manager.data['uri']))
+print("  with attribute 'displayName': {}".format(hypervisor_manager.data['displayName']))
 
-# Add another hypervisor manager and remove forcefully
-hypervisor_manager_added = oneview_client.hypervisor_managers.add(hypervisor_manager_information)
-oneview_client.hypervisor_managers.delete(hypervisor_manager_added, force=True)
-print("Successfully added and removed the hypervisor manager forcefully")
+# Delete the created hypervisor manager
+hypervisor_manager.delete()
+print("\nSuccessfully deleted hypervisor manager")