+++ CHANGELOG.md
@@ -68,4 +68,11 @@ Extends support of the SDK to OneView REST API version 1200 (OneView v5.00).
 - Storage volume template
 - Switch type
 - Uplink set
-- Hypervisor managers
+
+
+# 5.1.0 (unreleased version)
+#### Notes
+Provides SDK support to OneView REST API version 800, 1000 and 1200 (OneView v5.10).
+
+#### New Resource
+- Hypervisor Managers